Moved application inside web package. Upgraded Spring Boot to 2.1.3. Added banner.txt.
