+++ gradle.properties
@@ -1,2 +1,2 @@
-spring_boot_version=2.0.4.RELEASE
+spring_boot_version=2.1.3.RELEASE
 flyway_version=5.2.1
