+++ src/main/java/ua/lifebook/web/application/config/ConfigurationHolder.java
@@ -1,4 +1,4 @@
-package ua.lifebook.config;
+package ua.lifebook.web.application.config;
 
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigException;
@@ -12,8 +12,8 @@ import java.nio.file.Paths;
 
 public final class ConfigurationHolder {
     private static final Logger logger = LoggerFactory.getLogger(ConfigurationHolder.class);
-    public static final Path EXTERNAL_CONFIG_PATH = Paths.get("/etc/lbconf");
-    public static final String APP_NAME_PARAM = "app.name";
+    private static final Path EXTERNAL_CONFIG_PATH = Paths.get("/etc/lbconf");
+    private static final String APP_NAME_PARAM = "app.name";
 
     private final Config config = loadConfig();
 
@@ -22,18 +22,22 @@ public final class ConfigurationHolder {
         return tryLoadOverrides().withFallback(defaultConfig).resolve();
     }
 
+    public Config getConfig() {
+        return config;
+    }
+
     private Config tryLoadOverrides() {
         final Path configPath = getConfigFilePath();
 
-        logger.info("Loading configuration from {}", configPath.toAbsolutePath().toString());
+        logger.info("Loading configuration from {}", configPath.toAbsolutePath());
 
         return ConfigFactory.parseFileAnySyntax(configPath.toFile(), ConfigParseOptions.defaults().setAllowMissing(true));
     }
 
     /**
-     * Returns path for config file (without extension)
+     * Returns path for config file without extension
      */
-    public Path getConfigFilePath() {
+    private Path getConfigFilePath() {
         String appName;
         try {
             appName = System.getProperty(APP_NAME_PARAM, ConfigFactory.load().getString(APP_NAME_PARAM));
@@ -43,10 +47,5 @@ public final class ConfigurationHolder {
         }
 
         return EXTERNAL_CONFIG_PATH.resolve(appName);
-    }
-
-
-    public Config getConfig() {
-        return config;
     }
 }
