+++ src/main/resources/banner.txt
@@ -1,0 +1,5 @@
+   __ _  __      _                 _
+  / /(_)/ _| ___| |__   ___   ___ | | __
+ / / | | |_ / _ \ '_ \ / _ \ / _ \| |/ /
+/ /__| |  _|  __/ |_) | (_) | (_) |   <
+\____/_|_|  \___|_.__/ \___/ \___/|_|\_\
